+++ shell/installations/conda_install_for_gatk/install_exonerate_freebayes.sh
@@ -1,0 +1,10 @@
+eval "$(conda shell.bash hook)"
+
+conda activate gatk_modified
+conda install -y bioconda::exonerate
+conda deactivate
+
+conda create -y --name freebayes-env
+conda activate freebayes-env
+conda install -y bioconda::freebayes
+conda deactivate